Move hooks to root
