+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/AutocompleteDropdown.tsx
@@ -2,7 +2,7 @@
 import React, { ChangeEvent } from "react";
 import { t } from "i18next";
 
-import { useAutocompleteOptions } from "@clientComponents/form-builder/hooks/useAutocompleteOptions";
+import { useAutocompleteOptions } from "hooks/useAutocompleteOptions";
 
 interface DropdownOptionProps {
   label: string;
