+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/element-dialog/ElementFilters.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { Groups } from "@clientComponents/form-builder/hooks/useElementOptions";
+import { Groups } from "hooks/useElementOptions";
 import { cn } from "@lib/utils";
 import React, { useRef } from "react";
 import { RovingTabIndexProvider, useFocusEffect, useRovingTabIndex } from "react-roving-tabindex";
