+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/sub-elements/PanelHightlight.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React from "react";
 
-import { useIsWithin } from "@clientComponents/form-builder/hooks/useIsWithin";
+import { useIsWithin } from "hooks/useIsWithin";
 
 export const PanelHightLight = ({
   children,
