+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/sub-elements/SubElement.tsx
@@ -19,7 +19,7 @@ import {
   LockedBadge,
 } from "app/(gcforms)/[locale]/(form administration)/form-builder/components/shared";
 import { getQuestionNumber } from "@clientComponents/form-builder/util";
-import { useHandleAdd } from "@clientComponents/form-builder/hooks";
+import { useHandleAdd } from "hooks";
 
 export const SubElement = ({ item, elIndex, ...props }: { item: FormElement; elIndex: number }) => {
   const { t } = useTranslation("form-builder");
