+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/publish/Publish.tsx
@@ -3,11 +3,7 @@ import React, { useCallback, useState } from "react";
 import { useTranslation } from "@i18n/client";
 import { useRouter } from "next/navigation";
 import { useTemplateStore } from "@clientComponents/form-builder/store";
-import {
-  useTemplateApi,
-  useAllowPublish,
-  useRehydrate,
-} from "@clientComponents/form-builder/hooks";
+import { useTemplateApi, useAllowPublish, useRehydrate } from "hooks";
 import { CancelIcon, CircleCheckIcon, LockIcon } from "@serverComponents/icons";
 import { Button, Alert } from "@clientComponents/globals";
 import Link from "next/link";
