+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/SaveButton.tsx
@@ -5,7 +5,7 @@ import { cn } from "@lib/utils";
 
 import { Button, StyledLink } from "@clientComponents/globals";
 import { useTemplateStore } from "@clientComponents/form-builder/store";
-import { useTemplateStatus, useTemplateContext } from "@clientComponents/form-builder/hooks";
+import { useTemplateStatus, useTemplateContext } from "hooks";
 import { formatDateTime } from "@clientComponents/form-builder/util";
 import { SavedFailIcon, SavedCheckIcon } from "@serverComponents/icons";
 import { usePathname } from "next/navigation";
