+++ hooks/useAllowPublish.tsx
@@ -3,8 +3,12 @@ import { useCallback, useMemo } from "react";
 
 import { FormElement, FormElementTypes, FormProperties, PropertyChoices } from "@lib/types";
 import { useAccessControl } from "@lib/hooks";
-import { Description, publishRequiredFields, Title } from "../types";
-import { useTemplateStore } from "../store/useTemplateStore";
+import {
+  Description,
+  publishRequiredFields,
+  Title,
+} from "../components/clientComponents/form-builder/types";
+import { useTemplateStore } from "../components/clientComponents/form-builder/store/useTemplateStore";
 
 export class MissingTranslation extends Error {}
