+++ hooks/useTemplateContext.tsx
@@ -1,13 +1,16 @@
 "use client";
 import React, { createContext, useState, useContext, useRef, useCallback } from "react";
-import { useTemplateStore, useSubscibeToTemplateStore } from "../store";
+import {
+  useTemplateStore,
+  useSubscibeToTemplateStore,
+} from "../components/clientComponents/form-builder/store";
 import { useTemplateApi } from ".";
 import { useTranslation } from "@i18n/client";
 import { logMessage } from "@lib/logger";
 import { useSession } from "next-auth/react";
-import { toast } from "../../../../app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/Toast";
+import { toast } from "../app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/Toast";
 import { StyledLink } from "@clientComponents/globals";
-import { DownloadFileButton } from "../../../../app/(gcforms)/[locale]/(form administration)/form-builder/components/shared";
+import { DownloadFileButton } from "../app/(gcforms)/[locale]/(form administration)/form-builder/components/shared";
 
 interface TemplateApiType {
   error: string | null | undefined;
